[refactor_db] Relocate and rename database classes
