+++ asistente_ladm_col/config/config_db_supported.py
@@ -18,8 +18,8 @@
 """
 from qgis.PyQt.QtCore import QObject
 
-from ..db_support.pg_admin import PgAdmin
-from ..db_support.gpkg_admin import GpkgAdmin
+from ..lib.db.pg_factory import PgFactory
+from ..lib.db.gpkg_factory import GpkgFactory
 
 
 class ConfigDbSupported(QObject):
@@ -30,11 +30,11 @@ class ConfigDbSupported(QObject):
         self._init_db_items()
 
     def _init_db_items(self):
-        db_item = PgAdmin()
+        db_item = PgFactory()
         self._db_items[db_item.get_id()] = db_item
         self.id_default_db = db_item.get_id()
 
-        db_item = GpkgAdmin()
+        db_item = GpkgFactory()
         self._db_items[db_item.get_id()] = db_item
 
     def get_db_items(self):
