+++ asistente_ladm_col/gui/qgis_model_baker/dlg_import_schema.py
@@ -57,7 +57,7 @@ from ...utils.qt_utils import (Validators,
                                OverrideCursor)
 from ...resources_rc import *
 from ...config.config_db_supported import ConfigDbSupported
-from ...db_support.enum_action_type import EnumActionType
+from ...lib.db.enum_db_action_type import EnumDbActionType
 DIALOG_UI = get_ui_class('qgis_model_baker/dlg_import_schema.ui')
 
 class DialogImportSchema(QDialog, DIALOG_UI):
@@ -142,7 +142,7 @@ class DialogImportSchema(QDialog, DIALOG_UI):
     def show_settings(self):
         dlg = self.qgis_utils.get_settings_dialog()
         dlg.tabWidget.setCurrentIndex(SETTINGS_CONNECTION_TAB_INDEX)
-        dlg.set_action_type(EnumActionType.SCHEMA_IMPORT)
+        dlg.set_action_type(EnumDbActionType.SCHEMA_IMPORT)
         if dlg.exec_():
             self.db = dlg.get_db_connection()
             self._params = dlg.get_params()
